i59: Extract animations from useAnimation
