+++ src/res/functions/sheetAnimation.js
@@ -1,0 +1,4 @@
+export default () => {
+  const animation = 'foo';
+  return [animation];
+};
