+++ src/res/index.js
@@ -1,8 +1,18 @@
+// Resources
 export { default as COLORS } from './colors';
 export { default as TYPOGRAPHY } from './typography';
 export { default as SHAPES } from './shapes';
 export { default as STRINGS } from './strings';
 export { default as IMAGES } from './img';
 export { default as Icon } from './svg';
+
+// Functions
 export { default as normalizeStyles } from './functions/normalizeStyles';
 export { default as getThemeStyles } from './functions/getThemeStyles';
+
+// Animations
+export { default as bounceAnimation } from './functions/bounceAnimation';
+export { default as jiggleAnimation } from './functions/jiggleAnimation';
+export { default as menuAnimation } from './functions/menuAnimation';
+export { default as flipAnimation } from './functions/flipAnimation';
+export { default as sheetAnimation } from './functions/sheetAnimation';
